refactor: 프로젝트 등록 컴포넌트 섹션 분리
